requirements.txt
joblib==1.4.2
Flask==3.0.3
requests==2.32.3
reportlab==4.2.2
gunicorn==21.2.0
numpy==2.0.1
googletrans==4.0.0-rc1
python-dotenv==1.0.1
google-genai==0.6.0
